self review and better error messages
